requirements.txt
streamlit
requests
selenium
firebase-admin
google-cloud-firestore
